WIP: progress on seperation of concerns
