+++ brainles_aurora/inferer/inferer.py
@@ -8,7 +8,6 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Dict, List
 
-import nibabel as nib
 import numpy as np
 import torch
 from brainles_aurora.inferer import (
@@ -20,10 +19,8 @@ from brainles_aurora.inferer import (
     Output,
 )
 from brainles_aurora.inferer.data import DataHandler
+from brainles_aurora.inferer.model import ModelHandler
 from brainles_aurora.utils import download_model_weights, remove_path_suffixes
-from monai.inferers import SlidingWindowInferer
-from monai.networks.nets import BasicUNet
-from torch.utils.data import DataLoader
 
 logger = logging.getLogger(__name__)
 
@@ -41,13 +38,6 @@ class AbstractInferer(ABC):
         """
         self.config = config
         self._setup_logger()
-
-        # download weights if not present
-        self.lib_path: str = Path(os.path.dirname(os.path.abspath(__file__)))
-
-        self.model_weights_folder = self.lib_path.parent / "model_weights"
-        if not self.model_weights_folder.exists():
-            download_model_weights(target_folder=str(self.lib_path.parent))
 
     def _set_log_file(self, log_file: str | Path) -> None:
         """Set the log file for the inference run and remove the file handler from a potential previous run.
@@ -125,199 +115,8 @@ class AuroraInferer(AbstractInferer):
         super().__init__(config=config)
         logger.info(f"Initialized {self.__class__.__name__} with config: {self.config}")
         self.device = self._configure_device()
-
-        self.inference_mode = None
         self.data_handler = DataHandler(config=self.config)
-        # self.model_handler = ModelHandler(config=self.config)
-
-    def _get_model(self) -> torch.nn.Module:
-        """Get the Aurora model based on the inference mode.
-
-        Returns:
-            torch.nn.Module: Aurora model.
-        """
-
-        # init model
-        model = BasicUNet(
-            spatial_dims=3,
-            in_channels=len(self._get_not_none_files()),
-            out_channels=2,
-            features=(32, 32, 64, 128, 256, 32),
-            dropout=0.1,
-            act="mish",
-        )
-
-        # load weights
-        weights_path = os.path.join(
-            self.model_weights_folder,
-            self.inference_mode,
-            f"{self.config.model_selection}.tar",
-        )
-
-        if not os.path.exists(weights_path):
-            raise NotImplementedError(
-                f"No weights found for model {self.mode} and selection {self.config.model_selection}"
-            )
-
-        model = model.to(self.device)
-        checkpoint = torch.load(weights_path, map_location=self.device)
-
-        # The models were trained using DataParallel, hence we need to remove the 'module.' prefix
-        # for cpu inference to enable checkpoint loading (since DataParallel is not usable for CPU)
-        if self.device == torch.device("cpu"):
-            if "module." in list(checkpoint["model_state"].keys())[0]:
-                checkpoint["model_state"] = {
-                    k.replace("module.", ""): v
-                    for k, v in checkpoint["model_state"].items()
-                }
-        else:
-            model = torch.nn.parallel.DataParallel(model)
-
-        model.load_state_dict(checkpoint["model_state"])
-
-        return model
-
-    def _apply_test_time_augmentations(
-        self, outputs: torch.Tensor, data: Dict, inferer: SlidingWindowInferer
-    ) -> torch.Tensor:
-        """Apply test time augmentations to the model outputs.
-
-        Args:
-            outputs (torch.Tensor): Model outputs.
-            data (Dict): Input data.
-            inferer (SlidingWindowInferer): Sliding window inferer.
-
-        Returns:
-            torch.Tensor: Augmented model outputs.
-        """
-        n = 1.0
-        for _ in range(4):
-            # test time augmentations
-            _img = RandGaussianNoised(keys="images", prob=1.0, std=0.001)(data)[
-                "images"
-            ]
-
-            output = inferer(_img, self.model)
-            outputs += output
-            n += 1.0
-            for dims in [[2], [3]]:
-                flip_pred = inferer(torch.flip(_img, dims=dims), self.model)
-
-                output = torch.flip(flip_pred, dims=dims)
-                outputs += output
-                n += 1.0
-        outputs /= n
-        return outputs
-
-    def _save_as_nifti(self, postproc_data: Dict[str, np.ndarray]) -> None:
-        """Save post-processed data as NIFTI files.
-
-        Args:
-            postproc_data (Dict[str, np.ndarray]): Post-processed data.
-        """
-        # determine affine/ header
-        if self.data_handler.get_input_mode() == DataMode.NIFTI_FILE:
-            reference_file = self.data_handler.get_reference_nifti_file()
-            ref = nib.load(reference_file)
-            affine, header = ref.affine, ref.header
-        else:
-            logger.warning(
-                f"Writing NIFTI output after NumPy input, using default affine=np.eye(4) and header=None"
-            )
-            affine, header = np.eye(4), None
-
-        # save niftis
-        for key, data in postproc_data.items():
-            output_file = self.output_file_mapping[key]
-            if output_file:
-                output_image = nib.Nifti1Image(data, affine, header)
-                os.makedirs(os.path.dirname(output_file), exist_ok=True)
-                nib.save(output_image, output_file)
-                logger.info(f"Saved {key} to {output_file}")
-
-    def _post_process(
-        self, onehot_model_outputs_CHWD: torch.Tensor
-    ) -> Dict[str, np.ndarray]:
-        """Post-process the model outputs.
-
-        Args:
-            onehot_model_outputs_CHWD (torch.Tensor): One-hot encoded model outputs.
-
-        Returns:
-            Dict[str, np.ndarray]: Post-processed data.
-        """
-
-        # create segmentations
-        activated_outputs = (
-            (onehot_model_outputs_CHWD[0][:, :, :, :].sigmoid()).detach().cpu().numpy()
-        )
-        binarized_outputs = activated_outputs >= self.config.threshold
-        binarized_outputs = binarized_outputs.astype(np.uint8)
-
-        whole_metastasis = binarized_outputs[0]
-        enhancing_metastasis = binarized_outputs[1]
-
-        final_seg = whole_metastasis.copy()
-        final_seg[whole_metastasis == 1] = 1  # edema
-        final_seg[enhancing_metastasis == 1] = 2  # enhancing
-
-        whole_out = binarized_outputs[0]
-        enhancing_out = binarized_outputs[1]
-
-        # create output dict based on config
-        return {
-            Output.SEGMENTATION: final_seg,
-            Output.WHOLE_NETWORK: whole_out,
-            Output.METASTASIS_NETWORK: enhancing_out,
-        }
-
-    def _sliding_window_inference(
-        self, data_loader: DataLoader
-    ) -> Dict[str, np.ndarray]:
-        """Perform sliding window inference using monai.inferers.SlidingWindowInferer.
-
-        Args:
-            data_loader (DataLoader): Data loader.
-
-        Returns:
-            Dict[str, np.ndarray]: Post-processed data
-        """
-        inferer = SlidingWindowInferer(
-            roi_size=self.config.crop_size,  # = patch_size
-            sw_batch_size=self.config.sliding_window_batch_size,
-            sw_device=self.device,
-            device=self.device,
-            overlap=self.config.sliding_window_overlap,
-            mode="gaussian",
-            padding_mode="replicate",
-        )
-
-        with torch.no_grad():
-            self.model.eval()
-            self.model = self.model.to(self.device)
-            # currently always only 1 batch! TODO: potentialy add support to pass multiple image tuples at once?
-            for data in data_loader:
-                inputs = data["images"].to(self.device)
-
-                outputs = inferer(inputs, self.model)
-                if self.config.tta:
-                    logger.info("Applying test time augmentations")
-                    outputs = self._apply_test_time_augmentations(
-                        outputs, data, inferer
-                    )
-
-                logger.info("Post-processing data")
-                postprocessed_data = self._post_process(
-                    onehot_model_outputs_CHWD=outputs,
-                )
-
-                # save data to fie if paths are provided
-                if any(self.output_file_mapping.values()):
-                    logger.info("Saving post-processed data as NIFTI files")
-                    self._save_as_nifti(postproc_data=postprocessed_data)
-
-                logger.info("Returning post-processed data as Dict of Numpy arrays")
-                return postprocessed_data
+        self.model_handler = ModelHandler(config=self.config, device=self.device)
 
     def _configure_device(self) -> torch.device:
         """Configure the device for inference based on the specified config.device.
@@ -393,19 +192,13 @@ class AuroraInferer(AbstractInferer):
             images=validated_images
         )
 
+        self.model_handler.load_model(
+            inference_mode=determined_inference_mode,
+            num_input_modalities=self.data_handler.get_num_input_modalities(),
+        )
+
         logger.info("Setting up Dataloader")
         data_loader = self.data_handler.get_data_loader(images=validated_images)
-
-        if self.inference_mode != determined_inference_mode:
-            logger.info(
-                f"No loaded compatible model found (Switching from {self.inference_mode} to {determined_inference_mode}). Loading Model and weights"
-            )
-            self.inference_mode = determined_inference_mode
-            self.model = self._get_model()
-        else:
-            logger.info(
-                f"Same inference mode ({self.inference_mode}) as previous infer call. Re-using loaded model"
-            )
 
         # setup output file paths
         self.output_file_mapping = {
@@ -415,6 +208,6 @@ class AuroraInferer(AbstractInferer):
         }
 
         logger.info(f"Running inference on device := {self.device}")
-        out = self._sliding_window_inference(data_loader=data_loader)
+        out = self.model_handler
         logger.info(f"Finished inference {os.linesep}")
         return out
